+++ test/booleans.test.ts
@@ -1,7 +1,7 @@
 import { partition } from 'itertools';
 import { describe, expect, test } from 'vitest';
 
-import { boolean, numericBoolean, truthy } from '~/booleans';
+import { boolean, truthy } from '~/booleans';
 
 import { INPUTS } from './_fixtures';
 
@@ -37,24 +37,5 @@ describe('truthy', () => {
 
   test('invalid', () => {
     // truthy never fails
-  });
-});
-
-describe('numeric booleans', () => {
-  const decoder = numericBoolean;
-  const [okay, not_okay] = partition(INPUTS, (x) => Number.isFinite(x));
-
-  test('valid', () => {
-    expect(okay.length).not.toBe(0);
-    for (const value of okay) {
-      expect(decoder.verify(value)).toBe(!!value);
-    }
-  });
-
-  test('invalid', () => {
-    expect(not_okay.length).not.toBe(0);
-    for (const value of not_okay) {
-      expect(decoder.decode(value).ok).toBe(false);
-    }
   });
 });